Corrected CSV formatting errors, wrote new MappingStrategy for annotation-driven header position
The new MappingStrategy, SetOrderHeaderMappingStrategy<T>, takes a bean class that binds its fields both by name and position, and reads these annotations to determine where to put the columns. It also does not automatically sort or capitalize the headers like HeaderColumnNameMappingStrategy. Any fields which are missing either @CsvBindByPosition or one of @CsvBindByName and @CsvCustomBindByName will not be mapped.