Compute AST diff string for travis files, add to CSV output
Note: 14.6% of the sampled ML projects had .travis.yml files that were considered invalid by the Yaml tools we are using. More work will be needed to correct this to resolve it, though I am not sure if it is worth it or not. A comment in CommitAnalyzer.getYamlFileChangeAST() gives an example of what was considered invalid